% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_bundesland.R
\name{code_bundesland}
\alias{code_bundesland}
\title{Transforms the name of a Bundesland into the Bundesland ID}
\usage{
code_bundesland(string)
}
\arguments{
\item{x}{string (or vector of strings) to be converted into ID}
}
\value{
a character vector of valid Bundesland ID
}
\description{
Transforms the name of a Bundesland into the Bundesland ID
}
\details{
The first two digits of the 8 digit Amtlicher Gemeindeschlüssel (AGS) encode  
the Bundesland (Bundesland ID). This function converts the name of Bundesland into 
the Bundesland ID.
}
\examples{
 \dontrun{
 
  code_bundesland( "NRW" )

 }




}
\seealso{
\link{get_bundesland}
}
