% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_csv.R
\name{download_csv}
\alias{download_csv}
\title{Download the csv-file of a table}
\usage{
download_csv(tablename, genesis_db = "de")
}
\arguments{
\item{tablename}{name of the table to retrieve.}

\item{genesis_db}{name of the database (default: 'de').}
}
\description{
\code{download_csv()} downloads the csv for a table
}
\details{
Downloads the csv file to the working directory \code{getwd()}. 
This is an alternative approach to the retrieve_*() functions.
}
\examples{
 \dontrun{

 data <- download_csv("12411-0004.csv")

 }




}
\seealso{
\code{\link{read_header_genesis}} \code{\link{wiebaden}}.
}
