% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bundesland.R
\name{get_bundesland}
\alias{get_bundesland}
\title{Get the name or label of the Bundesland from the AGS}
\usage{
get_bundesland(num, render = "name", ags = FALSE)
}
\arguments{
\item{num}{a \code{vector} of codes to be converted}

\item{render}{Return \code{name} (default) or the \code{label}?}

\item{ags}{Only use the first two digits (default)?}
}
\value{
a character vector of names or labels.
}
\description{
Get the name or label of the Bundesland from the AGS
}
\details{
The first two digits of the 8 digit Amtlicher Gemeindeschlüssel (AGS) encode  
the Bundesland. This function converts them either to the name of the Bundesland 
(if \code{render="name"}) or to the two-character label (\code{render="label"}). 

\code{num} can either be the 8 digit AGS (set \code{ags=TRUE})
or the extracted two leading digits (\code{ags=FALSE}). If \code{num} is not 
a numeric vector, it is coerced to one.
}
\examples{
 \dontrun{
 
  get_bundesland( c("02", "03") )
  get_bundesland( c("03254021"), ags=TRUE )

 }




}
