% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieves Data from DESTATIS}
\usage{
retrieve_data(tablename, startyear = 1900, endyear = 2016, user = NULL,
  password = NULL, db = "regio")
}
\arguments{
\item{tablename}{name of the table to retrieve.}

\item{startyear}{only retrieve values for years equal or larger to \code{startyear}. Default: 1990.}

\item{endyear}{only retrieve values for years smaller or equal to \code{endyear}. Default: 2016.}

\item{user}{(required) user name.}

\item{password}{(required) password.}

\item{db}{select database. Default: regio (currently the only option).}
}
\value{
a \code{data.frame}. Value variables (_val) come with three additional variables (_qual, _lock, _err). The exact nature 
of these variables is unknown.
}
\description{
\code{retrieve_data} retrieves a single data table from DESTATIS
}
\details{
Use \code{retrieve_datalist} to find the \code{tablename} based on the series you are interested in. 

 Instead of setting user/password via the function, one can also set a vector with user/paswword in this form: 
 \code{destatis_user <- c(user="ABCDEF", password="XXXXX")}
}
\examples{
 \dontrun{
 abc. 
 }




}
\seealso{
\code{\link{retrieve_datalist}}
}

