% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_datalist.R
\name{retrieve_datalist}
\alias{retrieve_datalist}
\title{Retrieves Data List from DESTATIS}
\usage{
retrieve_datalist(user = NULL, tableseries, password = NULL, db = "regio")
}
\arguments{
\item{user}{(required) user name.}

\item{tableseries}{name of series for which tables should be retrieved.}

\item{password}{(required) password.}

\item{db}{select database. Default: regio (currently the only option).}
}
\value{
a \code{data.frame}
}
\description{
\code{retrieve_datalist} retrieves a list of available data tables in a series
}
\details{
abc
}
\examples{
 \dontrun{
 abc. 
 }




}
\seealso{
\code{\link{retrieve_data}}
}

