% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_metadata.R
\name{retrieve_metadata}
\alias{retrieve_metadata}
\title{Retrieves Meta Data from DESTATIS}
\usage{
retrieve_metadata(tablename, user = NULL, password = NULL, db = "regio")
}
\arguments{
\item{tablename}{name of the table to retrieve.}

\item{user}{(required) user name.}

\item{password}{(required) password.}

\item{db}{select database. Default: regio (currently the only option).}
}
\value{
a \code{data.frame}.
}
\description{
\code{retrieve_metadata} retrieves meta data.
}
\details{
Use \code{retrieve_datalist} to find the \code{tablename} based on the series you are interested in. 

 Instead of setting user/password via the function, one can also a vector with user/paswword in this form: 
 \code{destatis_user <- c(user="ABCDEF", password="XXXXX")}
}
\examples{
 \dontrun{
 abc. 
 }




}
\seealso{
\code{\link{retrieve_datalist}}
}

