% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_valuelabel.R
\name{retrieve_valuelabel}
\alias{retrieve_valuelabel}
\title{Retrieves Value Labels from DESTATIS}
\usage{
retrieve_valuelabel(variablename, valuelabel = "*", user = NULL,
  password = NULL, db = "regio")
}
\arguments{
\item{variablename}{name of the variable}

\item{valuelabel}{"*" (default) retrieves all value labels.}

\item{user}{(required) user name.}

\item{password}{(required) password.}

\item{db}{select database. Default: regio (currently the only option).}
}
\value{
a \code{data.frame}.
}
\description{
\code{retrieve_valuelabel} retrieves value labels for variable
}
\details{
Instead of setting user/password via the function, one can also a vector with user/paswword in this form: 
 \code{destatis_user <- c(user="ABCDEF", password="XXXXX")}
}
\examples{
 \dontrun{
 abc. 
 }




}
\seealso{
\code{\link{retrieve_datalist}}
}

