% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_credentials.R
\name{save_credentials}
\alias{save_credentials}
\title{Saves database credentials in home folder (only for Linux/Mac)}
\usage{
save_credentials(db, user, password, append = TRUE)
}
\arguments{
\item{db}{database name, either 'nrw', 'regio' or 'de'.}

\item{user}{user name.}

\item{password}{password.}

\item{append}{add credentials to file (default) or overwrite entire file?}
}
\value{

}
\description{
\code{save_credentials} saves a set of database credentials to \code{'~/.genesis.json'} to ease package usage.
}
\details{
Attention, user/password are stored in an unencrypted file in \code{'~/.genesis.json'}.
}
\seealso{
\code{\link{wiesbaden}}
}

