% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_login.R
\name{test_login}
\alias{test_login}
\title{Tests Login in GENESIS Databases}
\usage{
test_login(genesis = NULL, ...)
}
\arguments{
\item{genesis}{to authenticate a user and set the database (see below).}

\item{...}{other arguments send to the httr::GET request.}
}
\value{
a \code{string} with the server return message.
}
\description{
\code{test_login} tests if the login works.
}
\examples{

 \dontrun{

 test_login(genesis=c(db="regio") )

 }




}
