% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiesbaden-package.R
\docType{package}
\name{wiesbaden-package}
\alias{wiesbaden}
\alias{wiesbaden-package}
\title{Client to access the data from the Federal Statistical Office, Germany}
\description{
Data retrieval client for Federal Statistical Office of Germany
}
\details{
Feedback is very welcome!


\tabular{ll}{
	Package: \tab wiesbaden\cr
	Type: \tab Package\cr
	Version: \tab 0.0.0.9000\cr
	Date: \tab 2016-11-20\cr
	License: \tab  GPL-3\cr
	}
}
\author{
Moritz Marbach \email{marbach@ipz.uzh.ch}
}
\references{

}

