% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieves Data from DESTATIS}
\usage{
retrieve_data(tablename, startyear = 1900, endyear = 2016, user = NULL,
  password = NULL, db = "regio")
}
\arguments{
\item{startyear}{only retrieve values for years equal or larger to \code{startyear}. Default: 1990.}

\item{endyear}{only retrieve values for years smaller or equal to \code{endyear}. Default: 2016.}

\item{user}{(required) user name.}

\item{password}{(required) password.}

\item{db}{select database. Default: regio (currently the only option).}

\item{dataname}{name of the table to retrieve.}
}
\value{
a \code{data.frame}. Value variables (_val) come with three additional variables (_qual, _lock, _err). The exact nature 
of these variables is unknown.
}
\description{
\code{retrieve_data} retrieves a single data table from DESTATIS
}
\details{
abc
}
\examples{
 \dontrun{
 abc. 
 }




}
\seealso{
\code{\link{retrieve_datalist}}
}

