% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieves Data from GENESIS Databases}
\usage{
retrieve_data(tablename, startyear = 1900, endyear = 2016, user = NULL,
  password = NULL, db = "regio")
}
\arguments{
\item{tablename}{name of the table to retrieve.}

\item{startyear}{only retrieve values for years equal or larger to \code{startyear}. Default: 1990.}

\item{endyear}{only retrieve values for years smaller or equal to \code{endyear}. Default: 2016.}

\item{user}{user name (see below).}

\item{password}{password (see below).}

\item{db}{select database, default 'regio' (see below).}
}
\value{
a \code{data.frame}. Value variables (_val) come with three additional variables (_qual, _lock, _err). The exact nature 
of these variables is unknown.
}
\description{
\code{retrieve_data} retrieves a single data table.
}
\details{
Use \code{\link{retrieve_datalist}} to find the \code{tablename} based on the table series you are interested in. See the 
package description (\code{\link{wiebaden}}) for details about setting the login and database.
}
\examples{
 \dontrun{
 # Retrieve values for the table 14111KJ002 which contains the 
 # federal election results on the county level. 

 data <- retrieve_data(tablename="14111KJ002", user="ABCDEF", password="XXXXX", db="regio")
 }




}
\seealso{
\code{\link{retrieve_datalist}} \code{\link{wiebaden}}
}

