% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_datalist.R
\name{retrieve_datalist}
\alias{retrieve_datalist}
\title{Retrieves Data List from DESTATIS}
\usage{
retrieve_datalist(tableseries, user = NULL, password = NULL, db = "regio")
}
\arguments{
\item{tableseries}{name of series for which tables should be retrieved.}

\item{user}{(required) user name.}

\item{password}{(required) password.}

\item{db}{select database. Default: regio (currently the only option).}
}
\value{
a \code{data.frame}
}
\description{
\code{retrieve_datalist} retrieves a list of available data tables in a series
}
\details{
Instead of setting user/password via the function, one can also a vector with user/paswword in this form: 
 \code{destatis_user <- c(user="ABCDEF", password="XXXXX")}
}
\examples{
 \dontrun{
 abc. 
 }




}
\seealso{
\code{\link{retrieve_data}}
}

