% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_datalist.R
\name{retrieve_datalist}
\alias{retrieve_datalist}
\title{Retrieves List of Tables from GENESIS Databases}
\usage{
retrieve_datalist(tableseries, user = NULL, password = NULL, db = "regio")
}
\arguments{
\item{tableseries}{name of series for which tables should be retrieved.}

\item{user}{user name (see below).}

\item{password}{password (see below).}

\item{db}{select database, default 'regio' (see below).}
}
\value{
a \code{data.frame}
}
\description{
\code{retrieve_datalist} retrieves a list of available data tables in a series.
}
\details{
See the package description (\code{\link{wiebaden}}) for details about setting the login and database.
}
\examples{
 \dontrun{
 # Retrieves list of available tables for the table series 14111 
 # which contains the federal election results. 

 d <- retrieve_datalist(tableseries="14111")
 }




}
\seealso{
\code{\link{retrieve_data}} \code{\link{wiebaden}}
}

