% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_metadata.R
\name{retrieve_metadata}
\alias{retrieve_metadata}
\title{Retrieves Meta Data from DESTATIS}
\usage{
retrieve_metadata(tablename, user = NULL, password = NULL, db = "regio")
}
\arguments{
\item{user}{(required) user name.}

\item{password}{(required) password.}

\item{db}{select database. Default: regio (currently the only option).}

\item{dataname}{name of the table to retrieve.}
}
\value{
a \code{data.frame}.
}
\description{
\code{retrieve_metadata} retrieves meta data.
}
\details{
abc
}
\examples{
 \dontrun{
 abc. 
 }




}
\seealso{
\code{\link{retrieve_datalist}}
}

