% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_metadata.R
\name{retrieve_metadata}
\alias{retrieve_metadata}
\title{Retrieves Meta Data from GENESIS Databases}
\usage{
retrieve_metadata(tablename, user = NULL, password = NULL, db = "regio")
}
\arguments{
\item{tablename}{name of the table to retrieve.}

\item{user}{user name (see below).}

\item{password}{password (see below).}

\item{db}{select database, default 'regio' (see below).}
}
\value{
a \code{data.frame}.
}
\description{
\code{retrieve_metadata} retrieves meta data.
}
\details{
See the package description (\code{\link{wiebaden}}) for details about setting the login and database.
}
\examples{
 \dontrun{
 # Meta data contain the explanations to the variable names for the table
 # federal election results on the county level. 

 metadata <- retrieve_metadata(tablename="14111KJ002")
 }




}
\seealso{
\code{\link{wiebaden}}
}

