% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_valuelabel.R
\name{retrieve_valuelabel}
\alias{retrieve_valuelabel}
\title{Retrieves Value Labels from GENESIS Databases}
\usage{
retrieve_valuelabel(variablename, valuelabel = "*", user = NULL,
  password = NULL, db = "regio")
}
\arguments{
\item{variablename}{name of the variable}

\item{valuelabel}{"*" (default) retrieves all value labels.}

\item{user}{user name (see below).}

\item{password}{password (see below).}

\item{db}{select database, default 'regio' (see below).}
}
\value{
a \code{data.frame}.
}
\description{
\code{retrieve_valuelabel} retrieves value labels for variable
}
\details{
See the package description (\code{\link{wiebaden}}) for details about setting the login and database.
}
\examples{
 \dontrun{
 # Value labels contain for the variable 'NAT' in the table with the 
 # federal election results on the county level. 

 metadata <- retrieve_valuelabel(variablename="NAT")
 }




}
\seealso{
\code{\link{retrieve_datalist}} \code{\link{wiebaden}}
}

