% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_credentials.R
\name{save_credentials}
\alias{save_credentials}
\title{Saves database credentials}
\usage{
save_credentials(db, user, password)
}
\arguments{
\item{db}{database name, either 'nrw', 'regio', 'de' or 'bm'.}

\item{user}{your user name.}

\item{password}{your password.}
}
\description{
\code{save_credentials} saves a set of database credentials using the \code{keyring} package.
}
\details{
User/password are stored in Keychain on macOS, Credential Store on Windows or Secret Service API on Linux. 
 If a user/password pair for a database already exists, it is silently replaced with the new pair.  
 This function relies on the \code{\link{keyring}}  package.
}
\seealso{
\code{\link{wiesbaden}}, \code{\link{keyring}}
}
