% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiesbaden-package.R
\docType{package}
\name{wiesbaden-package}
\alias{wiesbaden-package}
\alias{wiesbaden}
\title{Client to access the data from the Federal Statistical Office, Germany}
\description{
Data retrieval client for Federal Statistical Office of Germany
}
\details{
To authenticate a user and set the database, supply a vector with user/password and database shortname in this form: 
\code{destatis_user <- c(user="your-username", password="your-password", db="database-shortname")}

Alternatively if you use Linux or Mac, place a json file named '.genesis.json' in your root directory (~). Use the 
\code{\link{save_credentials}} to generate this file. This only works with Linux/Mac!

Available databases are regionalstatistik.de (shortname: "regio"), landesdatenbank.nrw.de ("nrw"), 
www-genesis.destatis.de ("de") and bildungsmonitoring.de ("bm").
}
\author{
Moritz Marbach \email{moritz.marbach@gess.ethz.ch}
}
