% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiesbaden-package.R
\docType{package}
\name{wiesbaden-package}
\alias{wiesbaden-package}
\alias{wiesbaden}
\title{Client to access the data from the Federal Statistical Office, Germany}
\description{
Data retrieval client for Federal Statistical Office of Germany
}
\details{
To authenticate, supply a vector with your user name, password, and database 
shortcut ("regio", "de", "nrw", "bm") as an argument for the \code{genesis} 
parameter whenever you call a \code{retrieve_*} function:
\code{c(user="your-username", password="your-password", db="database-shortname")}

Alternatively, store the credentials on your computer using the \code{\link{save_credentials}} function. This function 
relies on the \code{\link{keyring}} package. 

Available databases are regionalstatistik.de (shortname: "regio"), landesdatenbank.nrw.de ("nrw"), 
www-genesis.destatis.de ("de") and bildungsmonitoring.de ("bm"), genesis.sachsen-anhalt.de ("st") and
statistikdaten.bayern.de ("by").
}
\author{
Moritz Marbach \email{moritz.marbach@tamu.edu}
}
